pyproject.toml
[project]
name = "mcp-http"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.13.0",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.10",
]
